Move the 'Display' struct to a separate module
This commit moves the 'Display' struct, used to draw emulator's output
on application's window, to a separate module named 'widgets'.
